there could a case where capacity is equal to string size, and cause invalid write(4 bytes); thus also the last 4-bytes are appended

change create_message interface; adding comments and descriptions to response message


remove cast to char


	modified:   ../src/rgw/rgw_rest_s3.cc
	modified:   ../src/rgw/rgw_rest_s3.h

+++ src/tools/rbd_wnbd/wnbd_handler.h
@@ -65,7 +65,7 @@ class WnbdHandler
 private:
   librbd::Image &image;
   std::string instance_name;
-  uint32_t block_count;
+  uint64_t block_count;
   uint32_t block_size;
   bool readonly;
   bool rbd_cache_enabled;
@@ -76,7 +76,7 @@ private:
 
 public:
   WnbdHandler(librbd::Image& _image, std::string _instance_name,
-              uint32_t _block_count, uint32_t _block_size,
+              uint64_t _block_count, uint32_t _block_size,
               bool _readonly, bool _rbd_cache_enabled,
               uint32_t _io_req_workers,
               uint32_t _io_reply_workers)
